added more detailed exception handling

+++ H_Plus_Sports/Controllers/CustomersController.cs
@@ -1,27 +1,30 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
+﻿using System.Linq;
 using System.Threading.Tasks;
-using H_Plus_Sports.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using H_Plus_Sports.Models;
+using System.Net;
 
-// For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
-
-namespace H_Plus_Sports.Controllers
+namespace HPlusSportsAPI.Controllers
 {
     [Produces("application/json")]
     [Route("api/Customers")]
     public class CustomersController : Controller
     {
         private readonly H_Plus_SportsContext _context;
+
         public CustomersController(H_Plus_SportsContext context)
         {
             _context = context;
         }
 
+        private bool CustomerExists(int id)
+        {
+            return _context.Customer.Any(e => e.CustomerId == id);
+        }
+
         [HttpGet]
+        [Produces(typeof(DbSet<Customer>))]
         public IActionResult GetCustomer()
         {
             var results = new ObjectResult(_context.Customer)
@@ -30,38 +33,91 @@ namespace H_Plus_Sports.Controllers
             };
 
             Request.HttpContext.Response.Headers.Add("X-Total-Count", _context.Customer.Count().ToString());
+
             return results;
         }
 
-        [HttpGet("{id}", Name = "GetCustomer")]
+        [HttpGet("{id}")]
+        [Produces(typeof(Customer))]
         public async Task<IActionResult> GetCustomer([FromRoute] int id)
         {
-            var customer = await _context.Customer.SingleOrDefaultAsync(m => m.CustomerId == id);
-            return Ok(customer);
-        }
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
 
-        [HttpPost]
-        public async Task<IActionResult> PostCustomer([FromBody] Customer customer)
-        {
-            _context.Customer.Add(customer);
-            await _context.SaveChangesAsync();
-            return CreatedAtAction("getCustomer", new { id = customer.CustomerId }, customer);
+            var customer = await _context.Customer.SingleOrDefaultAsync(m => m.CustomerId == id);
+
+            return customer == null ? NotFound() : (IActionResult)Ok(customer);
         }
 
         [HttpPut("{id}")]
+        [Produces(typeof(Customer))]
         public async Task<IActionResult> PutCustomer([FromRoute] int id, [FromBody] Customer customer)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            if (id != customer.CustomerId)
+            {
+                return BadRequest();
+            }
+
             _context.Entry(customer).State = EntityState.Modified;
+
+            try
+            {
+                await _context.SaveChangesAsync();
+                return Ok(customer);
+            }
+            catch (DbUpdateException)
+            {
+                if (!CustomerExists(id))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+            }
+        }
+
+        [HttpPost]
+        [Produces(typeof(Customer))]
+        public async Task<IActionResult> PostCustomer([FromBody] Customer customer)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            _context.Customer.Add(customer);
             await _context.SaveChangesAsync();
-            return Ok(customer);
+
+            return CreatedAtAction("GetCustomer", new { id = customer.CustomerId }, customer);
         }
 
         [HttpDelete("{id}")]
+        [Produces(typeof(Customer))]
         public async Task<IActionResult> DeleteCustomer([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var customer = await _context.Customer.SingleOrDefaultAsync(m => m.CustomerId == id);
+            if (customer == null)
+            {
+                return NotFound();
+            }
+
             _context.Customer.Remove(customer);
             await _context.SaveChangesAsync();
+
             return Ok(customer);
         }
     }
